Adjust the relationship between the exchangis-project module and exchangis-job module
