+++ exchangis-project/exchangis-project-entity/pom.xml
@@ -3,23 +3,30 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>exchangis</artifactId>
+        <artifactId>exchangis-project</artifactId>
         <groupId>com.webank.wedatasphere.exchangis</groupId>
         <version>1.1.2</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>exchangis-project</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>exchangis-project-server</module>
-        <module>exchangis-project-provider</module>
-        <module>exchangis-project-entity</module>
-    </modules>
+    <artifactId>exchangis-project-entity</artifactId>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
     </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.webank.wedatasphere.exchangis</groupId>
+            <artifactId>exchangis-dao</artifactId>
+            <version>1.1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-math3</artifactId>
+            <version>3.6.1</version>
+        </dependency>
+    </dependencies>
 
 </project>
