+++ exchangis-project/exchangis-project-server/pom.xml
@@ -18,14 +18,14 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.apache.linkis</groupId>
-            <artifactId>linkis-mybatis</artifactId>
-            <version>${linkis.version}</version>
+            <groupId>com.webank.wedatasphere.exchangis</groupId>
+            <artifactId>exchangis-project-provider</artifactId>
+            <version>1.1.2</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-math3</artifactId>
-            <version>3.6.1</version>
+            <groupId>com.webank.wedatasphere.exchangis</groupId>
+            <artifactId>exchangis-job-server</artifactId>
+            <version>1.1.2</version>
         </dependency>
         <dependency>
             <groupId>com.webank.wedatasphere.exchangis</groupId>
