+++ exchangis-project/exchangis-project-server/src/main/java/com/webank/wedatasphere/exchangis/project/server/restful/ExchangisProjectRestfulApi.java
@@ -7,16 +7,19 @@ import com.webank.wedatasphere.exchangis.common.enums.OperateTypeEnum;
 import com.webank.wedatasphere.exchangis.common.enums.TargetTypeEnum;
 import com.webank.wedatasphere.exchangis.common.pager.PageResult;
 import com.webank.wedatasphere.exchangis.common.validator.groups.UpdateGroup;
-import com.webank.wedatasphere.exchangis.project.server.domain.OperationType;
-import com.webank.wedatasphere.exchangis.project.server.entity.ExchangisProject;
-import com.webank.wedatasphere.exchangis.project.server.entity.ExchangisProjectUser;
+import com.webank.wedatasphere.exchangis.job.server.service.JobInfoService;
+import com.webank.wedatasphere.exchangis.job.vo.ExchangisJobQueryVo;
+import com.webank.wedatasphere.exchangis.job.vo.ExchangisJobVo;
+import com.webank.wedatasphere.exchangis.project.entity.domain.OperationType;
+import com.webank.wedatasphere.exchangis.project.entity.entity.ExchangisProject;
+import com.webank.wedatasphere.exchangis.project.entity.domain.ExchangisProjectUser;
 import com.webank.wedatasphere.exchangis.project.server.service.ProjectService;
 import com.webank.wedatasphere.exchangis.project.server.utils.ProjectAuthorityUtils;
 import com.webank.wedatasphere.exchangis.project.server.utils.ExchangisProjectConfiguration;
 import com.webank.wedatasphere.exchangis.project.server.utils.ExchangisProjectRestfulUtils;
-import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectInfo;
-import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectUserVo;
-import com.webank.wedatasphere.exchangis.project.server.vo.ProjectQueryVo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ExchangisProjectInfo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ExchangisProjectUserVo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ProjectQueryVo;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.math3.util.Pair;
 import org.apache.linkis.common.utils.JsonUtils;
@@ -48,6 +51,12 @@ public class ExchangisProjectRestfulApi {
      */
     @Resource
     private ProjectService projectService;
+
+    /**
+     * JobInfo service
+     */
+    @Resource
+    private JobInfoService jobInfoService;
 
     /**
      * Project query
@@ -208,6 +217,7 @@ public class ExchangisProjectRestfulApi {
                     .equalsIgnoreCase(domain)){
                 return Message.error("Cannot update the outer project (无法更新来自 " + domain + " 的外部项目)");
             }
+
             LOG.info("UpdateProject vo: {}, userName: {}", JsonUtils.jackson().writeValueAsString(projectVo), username);
             projectService.updateProject(projectVo, username);
             AuditLogUtils.printLog(oringinUser, username, TargetTypeEnum.PROJECT, projectVo.getId(), "Project name is: " + projectVo.getName(), OperateTypeEnum.UPDATE, request);
@@ -236,7 +246,16 @@ public class ExchangisProjectRestfulApi {
         try {
             ExchangisProjectInfo projectInfo = projectService.getProjectDetailById(id);
             if (!ProjectAuthorityUtils.hasProjectAuthority(username, projectInfo, OperationType.PROJECT_ALTER)) {
-                return Message.error("You have no permission to delete (删除项目失败)");
+                return Message.error("You have no permission to delete (没有权限删除项目)");
+            }
+
+            // 校验是否有任务
+            ExchangisJobQueryVo queryVo = new ExchangisJobQueryVo(
+                    id, null, null, null, null);
+            PageResult<ExchangisJobVo> exchangisJobVoPageResult = jobInfoService.queryJobList(queryVo);
+            if (Objects.nonNull(exchangisJobVoPageResult) && Objects.nonNull(exchangisJobVoPageResult.getList())
+                && exchangisJobVoPageResult.getList().size() > 0) {
+                return Message.error("Jobs already exist under this project and the project cannot be deleted (该项目下已存在子任务，无法删除)");
             }
 
             String domain = projectInfo.getDomain();
