+++ exchangis-project/exchangis-project-server/src/main/java/com/webank/wedatasphere/exchangis/project/server/service/impl/ProjectServiceImpl.java
@@ -10,14 +10,14 @@ import com.webank.wedatasphere.exchangis.job.api.ExchangisJobOpenService;
 import com.webank.wedatasphere.exchangis.job.domain.ExchangisJobEntity;
 import com.webank.wedatasphere.exchangis.job.exception.ExchangisJobException;
 import com.webank.wedatasphere.exchangis.job.vo.ExchangisJobQueryVo;
-import com.webank.wedatasphere.exchangis.project.server.entity.ExchangisProjectUser;
-import com.webank.wedatasphere.exchangis.project.server.mapper.ProjectMapper;
-import com.webank.wedatasphere.exchangis.project.server.entity.ExchangisProject;
-import com.webank.wedatasphere.exchangis.project.server.mapper.ProjectUserMapper;
+import com.webank.wedatasphere.exchangis.project.provider.mapper.ProjectMapper;
+import com.webank.wedatasphere.exchangis.project.provider.mapper.ProjectUserMapper;
 import com.webank.wedatasphere.exchangis.project.server.service.ProjectService;
-import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectInfo;
-import com.webank.wedatasphere.exchangis.project.server.vo.ExchangisProjectUserVo;
-import com.webank.wedatasphere.exchangis.project.server.vo.ProjectQueryVo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ExchangisProjectInfo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ExchangisProjectUserVo;
+import com.webank.wedatasphere.exchangis.project.entity.vo.ProjectQueryVo;
+import com.webank.wedatasphere.exchangis.project.entity.domain.ExchangisProjectUser;
+import com.webank.wedatasphere.exchangis.project.entity.entity.ExchangisProject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
